refactor(react): rename FlagSubscription to ConfigureAdapter
