+++ packages/react/modules/components/configure-adapter/index.js
@@ -1,0 +1,1 @@
+export default from './configure-adapter';
