--- packages/react/modules/components/flags-subscription/index.js
@@ -1,1 +1,0 @@
-export default from './flags-subscription';
