+++ packages/react/modules/components/index.js
@@ -3,4 +3,4 @@ export { default as SwitchFeature } from './switch-feature';
 export { default as injectFeatureToggle } from './inject-feature-toggle';
 export { default as injectFeatureToggles } from './inject-feature-toggles';
 export { default as branchOnFeatureToggle } from './branch-on-feature-toggle';
-export { default as FlagsSubscription } from './flags-subscription';
+export { default as ConfigureAdapter } from './configure-adapter';
